==> 1_Data_Processing.Rmd <==
## Data Processing
##### - for language enrollment rates prediction model
  
  
###### Doris Chen
###### 2024-04-05
  
***
  
**Load packages**

```{r, message = F}
library(openxlsx)
library(plyr)
library(tidyverse)
library(lubridate)
library(fs)
library(zoo)
```

#### Language Enrollments
To match the data availability from US Census, we will extract language enrollments data from the Modern Language Association database as follows: 

* Chinese/Japanese/Filipino: 1958-2021; 
* Korean: 1970-2021; 
* Hindi/Vietnamese: 1980-2021

```{r results='hide'}
# Load the data
MLA <- read.xlsx("https://www.mla.org/content/download/191323/file/Historical-language-enrollments-1958-2021.xlsx")

# Check column names
colnames(MLA)
```
```{r}
# Create a data frame for total enrollments for each census
totl_lan <- MLA %>% 
        filter(TERM == "Fall") %>% 
        dplyr::rename(Year = SRVY_YEAR) %>% 
        dplyr::group_by(Year) %>% 
        dplyr::summarise(Totl_Lan = sum(ALL.LEVEL.TOTAL, na.rm = TRUE)) %>% 
        dplyr::select(Year, Totl_Lan)

totl_lan <- totl_lan[!(totl_lan$Year == 2020), ]

```


```{r}
# Select needed languages and features
enrollments <- MLA %>% 
        filter(grepl("CHINESE|MANDARIN|JAPANESE|KOREAN|FILIPINO|HINDI|VIETNAMESE", LANGUAGE)) %>% 
        filter(TERM == "Fall") %>% 
        dplyr::rename(Year = SRVY_YEAR, Language = LANGUAGE) %>% 
        select(Year, Language, ALL.LEVEL.TOTAL) 
summary(enrollments)
```


```{r message = F}
# Change data types
enrollments <- enrollments %>% 
        mutate(Language = as.factor(Language))   

# Check all language course names
unique(enrollments$Language)

# Combine related language courses
enrollments$Language <- recode(enrollments$Language, "CHINESE" = "Chinese", "CHINESE, CLASSICAL" = "Chinese", 
                               "CHINESE, PRE-MODERN" = "Chinese",  "MANDARIN" = "Chinese", "JAPANESE" = "Japanese",
                               "JAPANESE, CLASSICAL" = "Japanese", "KOREAN" = "Korean",
                                "KOREAN, CLASSICAL" = "Korean", "HINDI-URDU" = "Hindi", 
                               "HINDI" = "Hindi", "VIETNAMESE" = "Vietnamese",
                               "FILIPINO/PILIPINO/TAGALOG" = "Filipino")     

# Calculate enrollments for each year-term-language
enrollments <- enrollments %>% 
        dplyr::group_by(Year, Language) %>% 
        dplyr::summarise(Lang_Enrol = sum(ALL.LEVEL.TOTAL, na.rm = TRUE))      

# Remove the enrollments of 2020 which was generated by a sample survey due to COVID-19
enrollments <- enrollments %>% 
        filter(Year != "2020")

# Remove 1958-1969 Korean enrollments and 1958-1979 Hindi/Vietnamese enrollments 
enrollments <- enrollments %>% 
        filter(!(Year < 1970 & Language == "Korean")) %>% 
        filter(!(Year < 1980 & Language %in% c("Hindi", "Vietnamese")))

summary(enrollments)
```
There is only 21 Filipino data while it supposed to be 22. From [MLA database](https://apps.mla.org/cgi-shl/docstudio/docs.pl?flsurvey_results), the enrollments of Filipino in 1961 is 0. We'll add the data into the data frame.

```{r}
enrollments[nrow(enrollments)+1, ] = list(1961, "Filipino", 0)
```

```{r}
# Join the enrollments and the totl_lan table to calculate the percentage of each language course enrollments over the total language course enrollments that year
enrollments <- left_join(enrollments, totl_lan, by = "Year")

enrollments <- enrollments %>% 
        mutate(Lang_Enrol_Pct = round(Lang_Enrol / Totl_Lan, 4)) %>% 
        select(-Totl_Lan)

```


We will interpolate the language enrollments data for the years between two nearest census (MLA conducted 26 census from 1958 to 2021) using spline.

```{r}
# Extend the data frame with the years needing interpolate
all_years <- seq(1958, 2021, by = 1)

ipt <- data.frame(Year = c(rep(all_years[!(all_years %in% enrollments[enrollments$Language == "Chinese",]$Year)], 3),
                           all_years[!(all_years %in% enrollments[enrollments$Language == "Korean",]$Year) & all_years > 1969],
                           rep(all_years[!(all_years %in% enrollments[enrollments$Language == "Vietnamese",]$Year) & all_years > 1979], 2)),
                  Language = c(rep("Chinese", 42), rep("Japanese", 42), rep("Filipino", 42),
                               rep("Korean", 37), rep("Vietnamese", 30), rep("Hindi", 30)),
                  Lang_Enrol = rep(NA, 223), Lang_Enrol_Pct = rep(NA, 223))

enrollments <- rbind(enrollments, ipt)

enrollments <- enrollments %>% 
        arrange(Language, Year)

```


```{r}
# Interpolating the missing data
for (lan in unique(enrollments$Language)){
       enrollments[enrollments$Language == lan, "Lang_Enrol"] = round(na.spline(enrollments[enrollments$Language == lan, "Lang_Enrol"]))
       enrollments[enrollments$Language == lan, "Lang_Enrol_Pct"] = round(na.spline(enrollments[enrollments$Language == lan, "Lang_Enrol_Pct"]), 4)
        
}

enrollments$Language <- as.factor(enrollments$Language)
summary(enrollments)
```


```{r}
ggplot(data = enrollments, aes(x = as.factor(Year), y = Lang_Enrol, group = Language, col = Language))+
        geom_line()+
        geom_point()+
        theme_classic()+
        theme(axis.text.x = element_text(angle = 90))+
        labs(title = "Language course enrollment trend", x = "Year", y = "Enrollment")
```


#### International trade 

```{r results = 'hide'}
# Load trade data 1985-2023 from US census
US_trade <- read.xlsx("https://www.census.gov/foreign-trade/balance/country.xlsx")
summary(US_trade)
```

```{r}
us_total <- read.xlsx("https://www.census.gov/foreign-trade/statistics/historical/gands.xlsx", startRow = 4, cols = c(1, 6, 9))
colnames(us_total) <- c("Year", "Exp", "Imp")
us_total <- us_total %>% 
        mutate(US_Total = Exp + Imp) %>% 
        mutate(Year = year(as.Date(Year, "%Y"))) %>% 
        select(Year, US_Total) 

us_total <- us_total[c(26:62), ]
summary(us_total)

```


```{r}
# Select needed countries, features and years
trade <- as.data.frame(US_trade[US_trade$CTYNAME %in% c("China", "Korea, South", "Japan", 
                                                        "India", "Vietnam", "Philippines"), ])
trade <- trade %>% 
        select(year, CTYNAME, IYR, EYR) %>% 
        filter(year <= 2021) %>% 
        mutate(Total_Trade = IYR + EYR) %>% 
        dplyr::rename(Language = CTYNAME, Year = year) %>% 
        mutate(Year = year(as.Date(Year, "%Y")), Language = as.factor(Language)) %>% 
        select(Year, Language, Total_Trade)

# Change country names to languages
trade$Language <- recode(trade$Language, "China" = "Chinese", "Korea, South" = "Korean", 
                         "Japan" = "Japanese", "Vietnam" = "Vietnamese", 
                         "India" = "Hindi", "Philippines" = "Filipino")

summary(trade)
```

From the summary, we found that Vietnam only got 30 data point, let's check which years are missing.
```{r}
trade[trade$Language == "Vietnamese",]$Year  
```

The Vietnam - US trade data starts from 1992, from [US Census](https://www.census.gov/foreign-trade/balance/c5520.html),  all figures are in millions of U.S. dollars, meaning before that the volume was less than 1 million, which we will treat it as 0.

```{r}
# Join the us_total table with the trade table, to calculate the trade_ratio
trade <- left_join(trade, us_total, by = "Year")

```

```{r}
trade <- trade %>% 
        mutate(Trade_Pct = round(Total_Trade / US_Total, 4)) %>% 
        select(Year, Language, Total_Trade, Trade_Pct)
```

Load trade data 1958--1984, which are derived from [Statistical Abstract of the United States](https://www.census.gov/library/publications/time-series/statistical_abstracts.html).

```{r message=FALSE}
# Combine trade data from 1958-2021
trade <- rbind(read_csv("Data/Trade 1958-1984.csv"), trade)
trade$Language <- as.factor(trade$Language)

summary(trade)
```


```{r}
ggplot(data = trade, aes(x = as.factor(as.character(Year)), y = Trade_Pct, group = Language, col = Language))+
        geom_point()+
        geom_line()+
        scale_y_continuous(labels = scales::percent)+
        theme_classic()+
        theme(axis.text.x = element_text(angle = 90))+
        labs(title = "US-foreign trade trend", x = "Year", y = "Percentage")
        
```


#### Total US university enrollments

```{r}
# Load the data
university <- read.xlsx("https://nces.ed.gov/programs/digest/d23/tables/xls/tabn303.10.xlsx", startRow = 5)
university <- university[c(11:dim(university)[1]), c(1,2)]
university <- as.data.frame(university[university[,1] %in% all_years,])
colnames(university) <- c("Year", "Total_Stud")
```

There are no enrollment data for 1958, 1960 and 1962 in the NCES dataset. We found the figures from a US census [report](https://www.census.gov/content/dam/Census/library/publications/1964/demo/p20-128.pdf#:~:text=The%20survey%20showed%20that%20about%2050.4%20mil%2D,period.%20The%20college%20and%20professional%20school%20level.) and impute the missing value.

```{r}
# Impute missing values
university[nrow(university)+1, ] <- c(1958, 3242000)
university[nrow(university)+1, ] <- c(1960, 3570000)
university[nrow(university)+1, ] <- c(1962, 4208000)

# Change the data type
university <- mutate(university, Year = year(as.Date(Year, "%Y")))
summary(university)
```

#### American Community Survey and US Census data

We downloaded the American Community Survey(ACS) S0201 table from [US census](https://data.census.gov/table/ACSSPP1Y2010.S0201?q=S0201&t=001:013:016:019:022:023:3767:3768:3782:3797:3804:029&g=010XX00US). The available years are 2010-2019 and 2021.

```{r}
# Load seven CSV files into a list
file_paths <- fs::dir_ls("Data/S0201")
ACS_file <- list()
for (i in seq_along(file_paths)){
        ACS_file[[i]] <- read_csv(
                file = file_paths[[i]], show_col_types = FALSE
        )
}
```


```{r}
# Clean the data 
acs_df <- list()
Year <- c(seq(2010, 2019, by = 1), 2021)

for (i in 1:length(ACS_file)){
        fil <- as.data.frame(ACS_file[i])
        fil$`Label..Grouping.` <- str_trim(fil$`Label..Grouping.`)
        fil <- fil[c(2, 97, 106, 165, 167, 168, 235, 272), ]
        colnames(fil) <- c("feature", "United_States", "Chinese", "Japanese", 
                                "Filipino", "Korean", "Hindi", "Vietnamese")
        
        us_fil <- fil[, c(1,2)]
        us_fil$United_States <- as.numeric(gsub("\\%", "", 
                                                gsub("\\,", "", us_fil$United_States)))
        
        fil <- fil[, c(1, 3:8)]
        
        rownames(fil) <- c(fil$feature)
        fil <- fil[, c(2:7)]
        fil <- t(fil)
        fil <- as.data.frame(fil)
        fil <- cbind(data.frame(Language = rownames(fil), Year = rep(Year[i], 6)), fil)
        rownames(fil) <- NULL
        cols_num <- 3:10
        fil[, cols_num] <- lapply(fil[, cols_num], function(fil){as.numeric(gsub("\\,|%", "", fil))})
        fil <- fil %>% 
                   mutate(`Bachelor's degree or higher` = `Bachelor's degree or higher`*0.01,
                          `Language other than English` = `Language other than English`*0.01,
                          `Speak English less than \"very well\"`=`Speak English less than \"very well\"`*0.01,
                          `All families` = `All families`*0.01) %>% 
                   mutate(HEdu_Pop = round(`Population 25 years and over` * `Bachelor's degree or higher`),
                          Lang_Pop = round(`Population 5 years and over` * `Language other than English`),
                          Eng_LesVw_Pop = round(`Speak English less than \"very well\"` * Lang_Pop)) %>% 
                   dplyr::rename(HEdu_Pct = `Bachelor's degree or higher`, Pop = `Total population`,
                          Med_Inc = `Median household income (dollars)`, Pvt_Pct = `All families`) %>% 
                   mutate(Pop_UsPct = round(Pop / us_fil[1,2], 4),
                          Lang_Pop_UsPct = round(Lang_Pop / us_fil[1,2], 4),
                          Eng_LesVw_Pop_UsPct = round(Eng_LesVw_Pop / us_fil[1,2], 4),
                          HEdu_UsPct = round(HEdu_Pop / round(us_fil[2,2] * us_fil[3,2] * 0.01), 4),
                          HEdu_Ratio = round(HEdu_Pct / (us_fil[3,2] * 0.01), 2),
                          Inc_Ratio = round(Med_Inc / us_fil[7,2], 2),
                          Pvt_Ratio = round(Pvt_Pct / (us_fil[8,2] * 0.01), 2),
                          ) %>% 
                   mutate(Language = as.factor(Language)) %>%    
                   select(Year, Language, Pop, Pop_UsPct, Lang_Pop, Lang_Pop_UsPct, Eng_LesVw_Pop, 
                          Eng_LesVw_Pop_UsPct, HEdu_Pop, HEdu_UsPct, HEdu_Pct, HEdu_Ratio,  
                          Med_Inc, Inc_Ratio, Pvt_Pct, Pvt_Ratio)
        acs_df[[i]] <- fil
             }

# Merge all data frames into one ACS2010-2021
ACS10_21 <- acs_df[[1]]
for (i in 2:11){
        ACS10_21 <- rbind(ACS10_21, acs_df[[i]])
}

summary(ACS10_21)
```
We derive US census 1950 -- 2000 data from the US decennial census report. These data share the same features as the ACS data above.

```{r}
# Load US census 1950-2000 data
US_census <- read_csv("Data/Census 1950-2000.csv")
summary(US_census)
US_census$Language <- as.factor(US_census$Language)

# Combine ACS10_21 and US_census data into one data frame
US_census <- rbind(US_census, ACS10_21)
```

However, there are still missing data for the years between the decennial census. As these are demographic data which usually change slowly and gradually, we will use cubic spline interpolation to estimate the missing data between two nearest decennial census. 


```{r}
# Extend the US_census data frame for interpolating data of the years between two decennial census
ipt <- data.frame(Year = c(rep(all_years[!(all_years %in% US_census$Year)],3),
                                all_years[all_years > 1969 & !(all_years %in% US_census$Year)],
                                rep(all_years[all_years > 1979 & !(all_years %in% US_census$Year)], 2)),
                       Language = c(rep("Chinese", 48), rep("Japanese", 48), rep("Filipino", 48), 
                                    rep("Korean", 37), rep("Hindi", 28), rep("Vietnamese", 28)))

for (i in 3:ncol(US_census)){
        ipt[, i] = rep(NA, nrow(ipt))
}

colnames(ipt) <- colnames(US_census)

US_census <- rbind(US_census, ipt)
US_census <- US_census %>% 
        arrange(Language, Year)

```

```{r}
# Interpolate missing data between two nearest decennial census using cubic spline interpolation
# The data related to Lang_Pop are only available since 1970; data related to Eng_LesVw are only available since 1980.

for (lan in unique(US_census$Language)){
     US_census[US_census$Language == lan & US_census$Year > 1969, "Lang_Pop"] = 
             na.spline(US_census[US_census$Language == lan & US_census$Year > 1969, "Lang_Pop"])
     US_census[US_census$Language == lan & US_census$Year > 1969, "Lang_Pop_UsPct"] = 
             na.spline(US_census[US_census$Language == lan & US_census$Year > 1969, "Lang_Pop_UsPct"])
     US_census[US_census$Language == lan & US_census$Year > 1979, "Eng_LesVw_Pop"] = 
             na.spline(US_census[US_census$Language == lan & US_census$Year > 1979, "Eng_LesVw_Pop"])
     US_census[US_census$Language == lan & US_census$Year > 1979, "Eng_LesVw_Pop_UsPct"] = 
             na.spline(US_census[US_census$Language == lan & US_census$Year > 1979, "Eng_LesVw_Pop_UsPct"])
        }

```


```{r}
# Interpolate the rest missing data 
col_nam <- colnames(US_census)[c(3:4, 9:16)]
lang <- unique(US_census$Language)

for (nam in col_nam){
     for (lan in lang){
             US_census[US_census$Language == lan, nam] = na.spline(US_census[US_census$Language == lan, nam])
     }  
}
```


```{r}
# Remove the unnecessary years
US_census <- US_census[US_census$Year %in% all_years, ]

# Round the figures to desired number of digits
for (nam in colnames(US_census)[3:16]){
        if (grepl("UsPct",  nam)){
                US_census[, nam] = round(US_census[, nam], 4)
        }
        else if (grepl("_Pct",  nam)){
                US_census[, nam] = round(US_census[, nam], 3)
        }
        else if (grepl("Ratio", nam)){
                US_census[, nam] = round(US_census[, nam], 2)
        } else {
                US_census[, nam] = round(US_census[, nam])
        }
}

summary(US_census)
```


#### International Students

```{r}
# Load the data
foreign_stud <- read.xlsx("https://opendoorsdata.org/?download=https://opendoorsdata.org/wp-content/uploads/2023/11/Census_All-Places-of-Origin_OD23.xlsx", startRow = 3)
```


```{r}
# Extract total international student data for each year 
world_total <- as.data.frame(foreign_stud[264, c(4:35)])

# Change the wide table to a long table
world_total <- pivot_longer(world_total, everything(), names_to = "Year", values_to = "world_total")

# Edit the year value and only keep the start year (fall)
world_total$Year <- substr(world_total$Year, 1, 4)

# Change the data type
world_total <- world_total %>% 
        mutate(Year = year(as.Date(Year, "%Y")), world_total = as.numeric(world_total))

summary(world_total)
```

```{r}
# Extract data related to the countries of interest
# Select needed rows and columns 
intl_stud <- as.data.frame(foreign_stud[foreign_stud$Place.of.Origin %in% c("China", "South Korea", "Japan", 
                                                                       "India", "Vietnam", "Philippines"), ])
intl_stud <- intl_stud[, -c(1, 3, 36:39)]

# change the wide table to a long table
intl_stud <- pivot_longer(intl_stud, !Place.of.Origin, names_to = "Year", values_to = "Intl_Stud")

# Edit the year value and only keep the start year (fall)
intl_stud$Year <- substr(intl_stud$Year, 1, 4)
summary(intl_stud)
```
```{r}
intl_stud$Intl_Stud <- gsub("\\-", "0", intl_stud$Intl_Stud)

intl_stud <- intl_stud %>% 
        dplyr::rename(Language = Place.of.Origin) %>% 
        mutate(Language = as.factor(Language), Year = year(as.Date(Year, "%Y")), 
               Intl_Stud = as.numeric(Intl_Stud)) %>% 
        select(Year, Language, Intl_Stud)

intl_stud$Language <- recode(intl_stud$Language, "China" = "Chinese", "South Korea" = "Korean", 
                              "Japan" = "Japanese", "Vietnam" = "Vietnamese", 
                              "India" = "Hindi", "Philippines" = "Filipino")
```


```{r}
# Join the world_total data with the intl_stud data to calculate the percentage of each country's international student over the total international student in US 

intl_stud <- left_join(intl_stud, world_total, by = "Year")
intl_stud <- intl_stud %>% 
        mutate(Intl_Stud_Pct = round(Intl_Stud / world_total, 3)) %>% 
        select(Year, Language, Intl_Stud, Intl_Stud_Pct)

summary(intl_stud)

```

The historical data from [Open Doors](https://opendoorsdata.org/data/international-students/all-places-of-origin/)  only provides figures every five years for the years before 2000. We found data for some years between 1960 and 1990 from the Open Doors annual reports available online and [NCES Digest of Education Statistics](https://nces.ed.gov/programs/digest/d95/dtab406.asp).

```{r}
# load international students data 1960-1990
stu60_98 <- read_csv("Data/Intl Students 1960-1998.csv")
intl_stud <- rbind(intl_stud, stu60_98)
intl_stud <- as.data.frame(intl_stud)
intl_stud <- arrange(intl_stud, Language, Year)

summary(intl_stud)
```

However, we cannot find data for all the years required for this project. Therefore, we will impute missing data using spline, similar to our approach with the census data.

```{r}
# Extend the data frame for missing years of international students data
ipt <- data.frame(
        Year = c(rep(all_years[!(all_years %in% intl_stud$Year)], 3),
                 all_years[all_years > 1969 & !(all_years %in% intl_stud$Year)],
                 rep(all_years[all_years > 1979 & !(all_years %in% intl_stud$Year)],2)),
        Language = c(rep("Chinese", length(all_years[!(all_years %in% intl_stud$Year)])),
                     rep("Japanese", length(all_years[!(all_years %in% intl_stud$Year)])),
                     rep("Filipino", length(all_years[!(all_years %in% intl_stud$Year)])),
                     rep("Korean", length(all_years[all_years > 1969 & !(all_years %in% intl_stud$Year)])),
                     rep("Hindi", length(all_years[all_years>1979 & !(all_years %in% intl_stud$Year)])),
                     rep("Vietnamese", length(all_years[all_years>1979 & !(all_years %in% intl_stud$Year)]))))


ipt[, ncol(ipt)+1] <- rep(NA, nrow(ipt))
ipt[, ncol(ipt)+1] <- rep(NA, nrow(ipt))
colnames(ipt)[3:4] <- c("Intl_Stud", "Intl_Stud_Pct")

intl_stud <- rbind(intl_stud, ipt)
intl_stud <- arrange(intl_stud, Language, Year)
```

We will interpolate missing data as following:
1. The number of Chinese student increased dramatically after the year 1979, so we will use linear interpolation via na.approx for Chinese students before year 1980 to avoid negative values generating by na.spine.
2. The rest of the missing data are interpolate by na.spline.

```{r}
intl_stud[intl_stud$Year < 1980 & intl_stud$Language == "Chinese", "Intl_Stud"] <-
        round(na.approx(intl_stud[intl_stud$Year < 1980 & intl_stud$Language == "Chinese", "Intl_Stud"]))

intl_stud[intl_stud$Year < 1980 & intl_stud$Language == "Chinese", "Intl_Stud_Pct"] <-
        round(na.approx(intl_stud[intl_stud$Year < 1980 & intl_stud$Language == "Chinese", "Intl_Stud_Pct"]), 3)

lang <- unique(intl_stud$Language)

for (lan in lang){
     intl_stud[intl_stud$Language == lan, "Intl_Stud"] = round(na.spline(intl_stud[intl_stud$Language == lan, "Intl_Stud"])) 
     intl_stud[intl_stud$Language == lan, "Intl_Stud_Pct"] = round(na.spline(intl_stud[intl_stud$Language == lan, "Intl_Stud_Pct"]), 3) 
        }

# Select years and remove unwanted data
intl_stud <- intl_stud[intl_stud$Year %in% all_years, ]
intl_stud <- intl_stud[!(intl_stud$Language == "Korean" & intl_stud$Year < 1970), ]
intl_stud <- intl_stud[!(intl_stud$Language %in% c("Hindi", "Vietnamese") & intl_stud$Year < 1980), ]

summary(intl_stud)
```


#### Party government
We derive the data of party government from [History, Art & Archives](https://history.house.gov/Institution/Presidents-Coinciding/Party-Government/)
```{r}
party_gov <- read_csv("Data/Party government.csv")
party_gov <- party_gov %>% 
        mutate(Presidency = as.factor(Presidency), Party_Gov = as.factor(Party_Gov))

summary(party_gov)
```

#### Combine all data frames

```{r}
all_data <- join_all(list(enrollments, US_census, intl_stud, trade), by = c("Year", "Language"), type = "left")
all_data <- join_all(list(all_data, university, party_gov), by = "Year", type = "left")

# Add a column "lan_UsPct" for the percentage of a language course enrollment over total US college enrollment
all_data <- all_data %>% 
        mutate(Lan_UsPct = round(Lang_Enrol/Total_Stud, 5))  

summary(all_data)
```


```{r}
# Save the data frame as csv file
write_csv(all_data, "Data/all_data.csv")
write_csv(totl_lan, "Data/All_Lang_Enrol.csv")
write_csv(university, "Data/UsCollege_Enrol.csv")
write_csv(enrollments, "Data/Asian_Lang_Enrol.csv")
write_csv(trade, "Data/trade.csv")
write_csv(US_census, "Data/US_census.csv")
```
